+++ yggdrasil_decision_forests/model/abstract_model.h
@@ -31,6 +31,7 @@
 #include "absl/status/statusor.h"
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
+#include "absl/types/span.h"
 #include "yggdrasil_decision_forests/dataset/data_spec.pb.h"
 #include "yggdrasil_decision_forests/dataset/example.pb.h"
 #include "yggdrasil_decision_forests/dataset/vertical_dataset.h"
@@ -655,6 +656,10 @@ void FloatToProtoPrediction(const std::vector<float>& src_prediction,
                             int example_idx, const proto::Task task,
                             int num_prediction_dimensions,
                             proto::Prediction* dst_prediction);
+
+// Converts a proto Prediction to a list of float predictions.
+void ProtoToFloatPrediction(const proto::Prediction& src_prediction,
+                            proto::Task task, absl::Span<float> dst_prediction);
 
 }  // namespace model
 }  // namespace yggdrasil_decision_forests